moved the part who run commands into a separated class: not all objects needs a full ArchInstall object. a CommandRunner is enoth
