+++ scripts/archinstall/arch/tools.py
@@ -1,5 +1,7 @@
 import os
+import subprocess
 from .mount import MountPoint
+from .archuser import ArchUser
 
 
 class Chroot():
@@ -68,3 +70,30 @@ class Cd():
 
     def __exit__(self, a, b, c):
         os.chdir(self.origin)
+
+
+class CommandRunner():
+    def __init__(self, mnt):
+        if not os.path.exists(mnt):
+            raise ValueError('invalid mount point you morron: ' + mnt)
+        self.mnt = mnt
+
+    def run(self, command, capture=False, critical=True, **kwargs):
+        if kwargs.get('debug_run'):
+            del(kwargs['debug_run'])
+            print('running', ' '.join(command), kwargs)
+        if capture:
+            return subprocess.check_output(command, **kwargs).decode('utf-8')
+        ret = subprocess.run(command, **kwargs)
+        if ret.returncode != 0 and critical:
+            raise CommandFail(command)
+
+    def run_in(self, command, user=None, **kwargs):
+        with ArchChroot(self.mnt):
+            if not user:
+                return self.run(command, **kwargs)
+            assert isinstance(user, ArchUser)
+            return self.run(command, preexec_fn=user.demote(), **kwargs)
+
+    def edit(self, filepath):
+        self.run_in(['vim', filepath])
